Added dialogue changes relating to BiomeName
Changed text dialogue mentioning biomes to work differently. Previously, it was "em o Deserto" and "de o Deserto". Now, with "n{BiomeName)" and "d{BiomeName}, it will appear as "no Deserto" and "do Deserto" respectively., which is the correct grammar. This might be a messy fix, but it works. Also fixed some instances where {BiomeName} would appear in the beggining of a sentence, which would result in it starting with an uncapitalized letter.